Updated BMP2TILE to version 0.4
BMP2TILE's version was updated to 0.4.

Some of the batch files had to be modified to compensate for
incompatibilities on BMP2TILE's command line syntax.

Tile loading routine had to be modified to take account the fact that
BMP2TILE ignores the tile offset when outputing the tiles in binary mode.